Proposed changed to allow multiple linked days.
This is a tentative change since the implications of this change
affect a wide-array of things in the application.

Reports rely on Linked Days, Various displays rely on it, calculations
etc. The proposed change has included a boatload of tests to allow try
to ensure that we don't step on any of the meaninful toes, the rest
can be discovered as and when. This sounds cavalier but I am not a
psychic.

+++ tracker/models.py
@@ -1140,7 +1140,7 @@ class TrackingEntry(models.Model):
     '''
 
     user = models.ForeignKey(Tbluser, related_name="user_tracking")
-    link = models.ManyToManyField("self", related_name="linked_entry",
+    link = models.ForeignKey("self", related_name="linked_entry",
                                 null=True, blank=True)
     entry_date = models.DateField()
     start_time = models.TimeField()
@@ -1230,12 +1230,12 @@ class TrackingEntry(models.Model):
         set the target link to a regular working day.
         '''
         if self.link:
-            if self.link.daytype == "LINKD":
+            if self.link.daytype == "LINKD" and \
+               len(TrackingEntry.objects.filter(link=self.link)) == 1:
                 self.link.delete()
             else:
-                self.link.link = None
-                self.link.save()
-
+                self.link = None
+                self.save()
     def breaktime(self):
         '''Returns the breaks entry of this tracking entry.'''
         return (dt.timedelta(hours=self.breaks.hour,
